Initial commit of arXiv ID library for Node.js
Note the regular expressions used in this version of the Identifiers
library differs slightly from the Ruby and PHP versions as JavaScript
does not support lookbehinds. Instead, we use capturing groups to ignore
the boundary before a match.

As this means we're matching a pattern multiple times with capturing
groups, we have to use `Regexp.prototype.exec()` in a loop until the
search is exhausted.